Remove opportunities for skew from PR 73
This was going to be a review of #73, but I was too slow.

The structure introduced to support the `--long` option means we need to
maintain the list of field names and corresponding data in two different
places, which opens up opportunities for skew (e.g, someone adds a new
field to the output without updating the behavior for `--long`).

This refactors the handling of `--long` so that it only ever augments the
data returned.