Fix type checks
The esiclient code contained a number of questionable type comparisons,
like this:

    if type(var) == "str":
        ...

But this is problematic, and can fail if what you've received is actually a
subclass of the target type. In general, we should always use
`isinstance()` for checking the type of a variable.

See also the documentation on flake8 rule E721 [1].

[1]: https://www.flake8rules.com/rules/E721.html